requirements.txt
#Requirements
black==22.3.0
mypy==0.931
pytest~=7.1.2
telegram~=0.0.1
Pillow >= 9.0
numpy~=1.22.2
scipy~=1.8.1
